requirements.txt
fastapi
uvicorn
langchain
pandas
numpy
matplotlib
python-dotenv
langchain-community
openai
langchain_openai